refactor: create app directory and move controllers & midlewares directory into app
