+++ app/controlllers/auth-controller.js
